+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/apideck-libraries/postman-to-k6/compare/v1.7.0...HEAD)
+## [Unreleased](https://github.com/apideck-libraries/postman-to-k6/compare/v1.8.0...HEAD)
+
+## [1.8.0] - 2020-09-18
 
 ### Added